refactor(client): renames all "repository" files to "graphql"
