+++ client/src/pages/sign-in/sign-in.service.tsx
@@ -1,6 +1,6 @@
 import { useMutation } from '@apollo/client';
 import { SignInMutation, SignInMutationVariables } from 'graphql/__generated__/graphql';
-import { SIGN_IN_MUTATION } from 'pages/sign-in/sign-in.repository';
+import { SIGN_IN_MUTATION } from 'pages/sign-in/sign-in.graphql';
 
 export const useSignIn = () => {
 	const [signIn, { loading }] = useMutation<SignInMutation, SignInMutationVariables>(
